+++ src/platform_impl/android/mod.rs
@@ -5,7 +5,7 @@ use std::{
     hash::Hash,
     sync::{
         atomic::{AtomicBool, Ordering},
-        mpsc, Arc,
+        mpsc, Arc, RwLock,
     },
     time::{Duration, Instant},
 };
@@ -14,6 +14,7 @@ use android_activity::input::{InputEvent, KeyAction, Keycode, MotionAction};
 use android_activity::{
     AndroidApp, AndroidAppWaker, ConfigurationRef, InputStatus, MainEvent, Rect,
 };
+use once_cell::sync::Lazy;
 use raw_window_handle::{
     AndroidDisplayHandle, HasRawWindowHandle, RawDisplayHandle, RawWindowHandle,
 };
@@ -26,6 +27,8 @@ use crate::{
     event_loop::{self, ControlFlow, EventLoopWindowTarget as RootELW},
     window::{self, CursorGrabMode, ResizeDirection, Theme, WindowButtons, WindowLevel},
 };
+
+static HAS_FOCUS: Lazy<RwLock<bool>> = Lazy::new(|| RwLock::new(true));
 
 fn ndk_keycode_to_virtualkeycode(keycode: Keycode) -> Option<event::VirtualKeyCode> {
     match keycode {
@@ -394,6 +397,7 @@ impl<T: 'static> EventLoop<T> {
                     warn!("TODO: find a way to notify application of content rect change");
                 }
                 MainEvent::GainedFocus => {
+                    *HAS_FOCUS.write().unwrap() = true;
                     sticky_exit_callback(
                         event::Event::WindowEvent {
                             window_id: window::WindowId(WindowId),
@@ -405,6 +409,7 @@ impl<T: 'static> EventLoop<T> {
                     );
                 }
                 MainEvent::LostFocus => {
+                    *HAS_FOCUS.write().unwrap() = false;
                     sticky_exit_callback(
                         event::Event::WindowEvent {
                             window_id: window::WindowId(WindowId),
@@ -947,6 +952,8 @@ impl Window {
 
     pub fn set_title(&self, _title: &str) {}
 
+    pub fn set_transparent(&self, _transparent: bool) {}
+
     pub fn set_visible(&self, _visibility: bool) {}
 
     pub fn is_visible(&self) -> Option<bool> {
@@ -966,6 +973,10 @@ impl Window {
     }
 
     pub fn set_minimized(&self, _minimized: bool) {}
+
+    pub fn is_minimized(&self) -> Option<bool> {
+        None
+    }
 
     pub fn set_maximized(&self, _maximized: bool) {}
 
@@ -1060,6 +1071,10 @@ impl Window {
 
     pub fn theme(&self) -> Option<Theme> {
         None
+    }
+
+    pub fn has_focus(&self) -> bool {
+        *HAS_FOCUS.read().unwrap()
     }
 
     pub fn title(&self) -> String {