+++ src/platform_impl/linux/x11/event_processor.rs
@@ -930,6 +930,10 @@ impl<T: 'static> EventProcessor<T> {
                             let window_id = mkwid(xev.event);
                             let position = PhysicalPosition::new(xev.event_x, xev.event_y);
 
+                            if let Some(window) = self.with_window(xev.event, Arc::clone) {
+                                window.shared_state_lock().has_focus = true;
+                            }
+
                             callback(Event::WindowEvent {
                                 window_id,
                                 event: Focused(true),
@@ -1001,6 +1005,10 @@ impl<T: 'static> EventProcessor<T> {
                                 window_id,
                                 event: WindowEvent::ModifiersChanged(ModifiersState::empty()),
                             });
+
+                            if let Some(window) = self.with_window(xev.event, Arc::clone) {
+                                window.shared_state_lock().has_focus = false;
+                            }
 
                             callback(Event::WindowEvent {
                                 window_id,